feat: implemented the stockGestion and the getAvalaible quantity in ingredient

+++ src/main/java/org/radiant_wizard/db/migration/dish_table_creation.sql
@@ -1,5 +1,5 @@
 create table dishes(
-    dish_id int primary key,
+    dish_id BIGSERIAL primary key,
     dish_name varchar(50) not null,
     dish_price NUMERIC(10,2)
 );
