+++ src/main/java/org/radiant_wizard/db/migration/ingredients_table_creation.sql
@@ -1,7 +1,7 @@
 create type measurement_unit as ENUM('G', 'L', 'U');
 
 create table ingredients(
-    ingredient_id int primary key,
+    ingredient_id BIGSERIAL primary key,
     ingredient_name varchar(50) not null,
     last_modification timestamp without time zone,
     unit_price NUMERIC(10,2),
